Move Recorder to its own file
