+++ src/background/background-script.js
@@ -2,6 +2,7 @@ import compat from '../shared/compat.js';
 import { product } from '../shared/func.js';
 import TLTranslationHelper from './TLTranslationHelper.js';
 import WASMTranslationHelper from './WASMTranslationHelper.js';
+import Recorder from './Recorder.js';
 
 
 function isSameDomain(url1, url2) {
@@ -301,74 +302,6 @@ function updateMenuItems({data, target: {state}}) {
         visible: state.models?.some(({from, to}) => from === state.from && to === state.to)
               && state.models?.some(({from, to}) => from === state.to && to === state.from)
     });
-}
-
-/**
- * A record can be used to record all translation messages send to the
- * translation backend. Useful for debugging & benchmarking.
- */
-class Recorder {
-    #pages;
-
-    constructor() {
-        this.#pages = new Map();
-    }
-
-    record({from, text, session: {url}}) {
-        // Unique per page url
-        if (!this.#pages.has(url))
-            this.#pages.set(url, {
-                url,
-                from,
-                texts: [],
-            });
-
-        // TODO: we assume everything is HTML or not, `html` is ignored.
-        this.#pages.get(url).texts.push(text);
-    }
-
-    get size() {
-        return this.#pages.size;
-    }
-
-    clear() {
-        this.#pages.clear();
-    }
-
-    exportAXML() {
-        const root = document.implementation.createDocument('', '', null);
-        const dataset = root.createElement('dataset');
-
-        this.#pages.forEach(page => {
-            const doc = root.createElement('doc');
-            doc.setAttribute('origlang', page.from);
-            doc.setAttribute('href', page.url);
-
-            const src = root.createElement('src');
-            src.setAttribute('lang', page.from);
-
-            page.texts.forEach((text, i) => {
-                const p = root.createElement('p');
-                
-                const seg = root.createElement('seg');
-                seg.setAttribute('id', i + 1);
-
-                seg.appendChild(root.createTextNode(text));
-                p.appendChild(seg);
-
-                src.appendChild(p);
-            });
-
-            doc.appendChild(src);
-            dataset.appendChild(doc);
-        });
-
-        root.appendChild(dataset);
-
-        const serializer = new XMLSerializer();
-        const xml = serializer.serializeToString(root);
-        return new Blob([xml], {type: 'application/xml'});
-    }
 }
 
 // Supported translation providers
